aws-as-activate: generate setup and teardown functionality
You can now run this an have the core of the framework installed in your
current shell environment:

    $ eval "$(./aws-as-activate -s)"

OR (equivalently for non-csh users):

    $ eval "$(./aws-as-activate)"

That will install the cleanup wiring and the user-facing
'aws-as-deactivate' function, which can be invoked to tear it all down.

It will set the "private global" variable __T_V_AWS_AS_PROFILE to the
value '-', which is our shorthand for "no AWS profile is active". That
variable is intended to get updated as the user moves between AWS
profiles. Each time that happens, this function should be invoked:

    __t_f_aws_as_maybe_update_ps1()

It uses the above variable to set a new value for $PS1, based on the
aws-as prefix "(${aws_profile}) " plus whatever the user had $PS1 set to
previously.

The "installed" shell functions will also remember your current PS1
setting, if necessary, and restore it when you invoke
'aws-as-deactivate'.