update file: NEWS
This just captures a brain dump about how the aws-cli 'configure'
subcommand can be used to set the 'aws_as_mfa_serial', and the fact that
we are laying claim to 'aws_as_' as a "namespace" with which to prefix
variables specific to 'aws-as'.

This catpures the rationale, but the 'NEWS' file is not really the right
place for this type of information. Cleaned up pieces of it should go
into the README.md and/or a "quick start" section the 'INSTALL' file.