aws-as: create ~/.aws-as.d and cache subdir upon activation, if needed
We have some autoconfiscated mojo for explicitly invoking chmod, but the
shell snippet that would use it is commented-out. My machine has 'mkdir'
from the GNU coreutils, which allows us to provide the '--mode' opt. I'm
not sure how portable that is, though, and I don't have time to look
into it at the moment.