aws-as: working impl. of aws-cli cache filename generation
The SHA-1 sum of the aws-cli cache files is based on a set of sorted key names
and their associated values. Because the serialized JSON string is important
to the hashcode, we here embed a Python 3 one-liner to calculate such names
for profiles associated with Roles; the hashcode is specific to the output of
the Python 'json' dumps method. Rather than try to emulate it, it's cheap
enough to just use the Python version directly.