fixes #6: new opt: '-C' (--clear-creds); '-' no longer clears
The way to clear the 'AWS_*' environment variables (if any) from the
environment of the current shell process is now via the new option:

    '-C' (--clear-creds)

Prior to this change, the way to perform that operation was to specify
the single character '-' as the aws-cli profile name command line
param. We want to use that syntax for something else (see issue #5), so
here we provide a different means of achieving the same end.