fixes #17: aws-as: extract function for parsing creds cache file content
All parsing of cached credentials file content is now performed by the
new funtion:

    __t_f_aws_as_parse_creds_from_cache_file ()

which was extracted from the inline logic used in a couple of different
places. See the comment in the source code for the function's calling
convention, which uses global variables named by the caller to "return"
the parsed values individually.

Because the data stored in those global variables is sensitive, a helper
function is provided, used both for initialization and cleanup (so we do
not leave the global variable set once the caller has obtained the
parsed data values):

    __t_v_aws_as_parsed_creds_reset ()