fixes #19: aws-as: obtain parsed creds through the cache, always
When fresh user (sts:GetSessionToken) or role (sts:AssumeRole) creds are
obtained from the AWS API (via aws-cli), the code used to parse the
creds out of the in-memory JSON response and use them directly, and
later write the response JSON to the cache file. That meant the
mechanism used to establish "fresh" creds was different from how the
creds would be read from the cache later on. That approach left room (in
the presence of bugs) for the thing in effect upon first use being
different from the thing that will be used in the future from the cache.

This change causes the response JSON to get written to the cache file
first, and the parsed credential values are then read back in through
the cache. The only credentials used are the ones read from our cache
interface. This will force problems in this area to the surface earlier,
and simplifies the implementation because we do the same thing in the
same way consistently with regard to obtaining credentials.

It also has the side effect of detecting (major) clock skew on the
machine earlier, which will force it to be dealt with. The previous
approach could leave the cache simply not working in such a scenario; in
some circumstance there would simply never be a cache hit for
non-expired credentials (if they always looked expired). With the new
approach, the "expired?" check is performed when the creds are read from
the cache; if the just-written fresh credentials are detected to be
expired, the program exits with an error hinting that the system clock
may need adjustment.