[4 of 8] issue #25: aws-as: add __t_f_aws_as_try_set_profile_as2()
This function implements the caching of the aws-cli creds to
aws-vault. It is invoked from:

    __t_f_aws_as_main()

instead of the __t_f_aws_as_try_set_profile_as() function. This shows
the feature working without (yet) gutting the existing
implementation. Future commits will remove the older implementation, and
replace it with this new one, at which point it will be renamed to get
rid of the '2' suffix.