[6 of 8] issue #25: aws-as: rip-out home-grown, file-based caching impl.
Major surgery: rip out our file-based caching implementation; it is no
longer needed. Bump future release number from 0.1.1-snapshot to
0.2.0-snapshot to reflect the severity of the changes.

In the generated 'aws-as' program, the following functions are removed:

    __t_f_aws_as_parse_creds_from_cache_file()
    __t_v_aws_as_parsed_creds_reset()
    __t_f_aws_as_try_set_profile_as_for_user()
    __t_f_aws_as_try_set_profile_as_for_role()
    __t_f_aws_as_try_set_profile_as()

The in-memory profile-name-to-User-ARN cache is removed:

    __T_V_AWS_AS_PROF_TO_UARN

Also removed is support for the ~/.aws-as.d/cache/ subdirectory. Delete
it if you have one:

    $ rm -fr ~/.aws-as.d/cache