[3 of 4] issue #30: aws-as: add '-c' (--clear-env) option
A new '-c' ("little cee") option has been added to clear the AWS_*
credential variables (if set) from the current environment.

This is the behavior previously performed by '-C' ("capital cee"), but
the meaning of that option will be expanded in an upcoming commit.

Rename function:

    __t_f_aws_as_clear_out_creds()

to:
    __t_f_aws_as_clear_env_of_creds()

to reflect its more targetted effect.