[4 of 4] fixes #30: aws-as: '-c' (--clear-creds) now clears from aws-vault
The '-C' (--clear-creds) option can now be thought of as a stronger form
of '-c'. It implies the new '-c' (--clear-env) opt, but also requests
that 'aws-vault' clear the temporary credentials from its cache for the
in-effect aws-cli profile.

It also implies '--none', which unsets the in-effect profile (because
using the profile again may require re-authentication -- it no longer
has an active "session" with cached creds in the aws-vault cache).