[1 of 3] issue #26: aws-as: add '-H' (--hoist) option
Two generated shell scripts are involved. The first, the "subshell
script", is supplied to an 'aws-vault' procress to be invoked by a
grandchild bash process. It plucks the 'AWS_*' credential variables out
of its environment and emits the second script (the "response script")
on stdout that sets the variables (basically a handful of 'declare -g
NAME=VALUE' statements).

The 'aws-as' command function running in the user's current shell eval's
the response script, thereby using the values obtained from the
subprocess to set the AWS_* credential variables in the current shell.