fixes #27: add cfg file knobs to support init debugging, auto-hoisting
  * aws_as_cfg_debug_init (true|false)

    If enabled, will emit debugging messages related to the loading to
    the config file variables.

  * aws_as_cfg_auto_hoist (true|false)

    If enabled, will cause aws-as to automatically "hoist" the AWS_*
    credential variables into the current environment when switching to
    any aws-cli profile.

    The hoisted variables are not, by default, exported.

This chanset also provide support for allowing the config file variables
to be specified on the 'aws-as-activate -s' command line. When specified
in this way, they override whatever values are in the config file, which
can be handy for one-off invocations (but see caveat below).

Instead of doing this:

    $ eval "$(aws-as-activate -s)"

you can do something like this:

    $ eval "$(printf '%s;\n' 'aws_as_cfg_debug_init=true' && aws-as-activate -s)"

or even this:

    $ eval "$(printf '%s;\n' 'set -x' 'aws_as_cfg_debug_init=true' 'set
    +x' && aws-as-activate -s)"

It also works to just set the variable in the current shell environment,
too. They are evaluated in the 'eval' context of the current shell at
activation time (and their original values remembered and restored at
deactivation time).

CAVEAT: Any variables set as described above will look to 'aws-as' to be
        pre-existing in the environment, so will be restored to the
        original setting as specified on the command line at
        deactivation time; they will still be present in the environment
        after you invoke aws-as-deactivate.