fixes #5: 'aws-as -' now recognized to mean "switch to the other profile"
This allows even faster switching back and forth between two
profiles. The behavior is intended to be analogous to the 'cd -' command
in your shell to hop back and forth between two directories.

Upon activation, there will not yet be an "old profile" to hop back to,
so using this syntax in that scenario will yield the message (on
stderr):

    aws-as: old profile not set

The same is actually true until you switch to a second profile within a
session, for obvious reasons.