fixes #4: aws-as-activate: generate 'aagg' command
Add 'aagg' to the list of user-facing "command functions" (shell
functions that look and feel a lot like a normal command line program)
emitted by:

    aws-as-activate -s

The job of 'aagg' is to run a user-provided command in the content of
the in-effect profile:

    aagg YOUR_COMMAND [ARG...]

The 'aagg' command will build the approprate 'aws-vault' command line to
use the (already authenticated) aws-cli credentials for the in-effect
profile. So the above would cause an invocation like this:

    /path/to/aws-vault [OPT...] exec IN_EFFECT_PROFILE_NAME -- YOUR_COMMAND [ARG...]

If there is no in-effect profile, then the user will see this:

    (_) $ aagg aws sts get-caller-identity
    aagg (error): no in-effect profile; see aws-as(1) to set it up

As you would expect, the 'aagg' function and all of it's supporting
machinery get deleted from the current shell process when the user
invokes 'aws-as-deactivate'.

About the name:
---------------
    aagg == (A)WS-(a)-(g)o-(g)o
    Read as: "AWS-a-go-go"
    Dumb name; is easy to type; is grep-friendly.

In this case, the "easy to type" part is essential because the user
needs to prefix it to every "real" command to be run, and it is expected
that it will need to be typed quite frequently.