update file: stylesheets/styles.css
This changeset represents the CSS changes as I found them as I sat down
to work this morning. Judging from the timestamp on the file, the last
changes were three days ago (at 2020-09-14 18:37:23 EDT).

The goal of the changes was to make better use of the screen realestate
to allow more content to be displayed without scrolling. In particular,
the code snippet sections previously had significant side scrolling
which has now been all but eliminated.

The main change is that the prose text is both justified and limited in
width (so the columns do not get too wide). The prose width is now
constrained more than the code snippet width, which produces a more
comfortable reading experience (code snippets are just a different
beast, so different presentation rules should apply).

I'm still not happy with it. The layout only really works for a ~900px
wide area. It kinda scales down okay (enough), but doesn't scale up
well. In particular, now matter how wide your screen is the code areas
will only take up so much space before a side-scrolling scrollbar is
presented.