update file: stylesheets/styles.css: introduce class "article-content"
The CSS class "article content" allows us to style container
elements (such as 'div') so that their child elements will inherit their
settings.

This is handy now that we've adjusted things to allow the 'pre' sections
to be wider than the prose in the article sections (which were capped
earlier; see commit c9410485894cf774f04f85c67c7cf3a9dcec23c2). In
particular, it allows us to constrain the width of paragraph text that
appears in list items (both 'ul' and 'ol') to not exceed the width of
the text in standalone paragraphs.

Explanation:
------------
Because HTML does not support block-level elements (such as '<ul>' and
'<ol>' inside of paragraph ('<p>') tags, the list items are logically
outside of the 'max-width' CSS styling that was already present for
paragraphs. However, when we add a <p> element as a child of a <li> tag,
then the paragraph width from the CSS does apply, but it's left side
starting point is indented due to being inside of a list; that means
it's right side is "futher out" than the right side boundary of the
normal article paragraph text. That situation causes the list element
paragraphs to render in the same width as the article paragraphs, but
the block of space they occupy is shifted to the right.

Semantically, the lists in our document tend to be part of a paragraph,
but HTML does not allow that semantic to be expressed directly -- well,
at least not within the syntactic construct of the paragraph, which is
how we think about it.

Instead, we use a '<div>' tag to contain the logical paragraph and it's
other block-level elements that we want to work within the same overall
width. In this case, we have something like this:

    <div class=".article-content">
        <p>...</p>
        <ol>
            <li><p>...</p></li>
            <li><p>...</p></li>
            <li><p>...</p></li>
        </ol>
        <p>...</p>
    </div>

Note that all of that is logically a single paragraph that contains a
list, where each of the list elements contains a relatively long prose
description.

Note that the style settings we apply to the "article-content" class are
the same as those applied to the 'p' element (the "max-width" setting,
in particular). Therefore the above approach of explicitly applying the
class is only needed when the "logical paragraph" contains one or more
block-level elements whose width needs to be constrained.